% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsgl.R
\docType{data}
\name{AirlineTicketPrices}
\alias{AirlineTicketPrices}
\title{Airline Ticket Prices.}
\format{A design matrix and a response matrix
\describe{
  \item{X}{design matrix}
  \item{Y}{response matrix}
}}
\description{
Airline Ticket Prices.
}
\keyword{data}

