% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.R
\name{features_stat.lsgl}
\alias{features_stat.lsgl}
\title{Extract feature statistics}
\usage{
\method{features_stat}{lsgl}(object, ...)
}
\arguments{
\item{object}{a lsgl object}

\item{...}{ignored}
}
\value{
a vector of length \code{nmod(x)} or a matrix containing the number of nonzero features (or group) of the models.
}
\description{
Extracts the number of nonzero features (or group) in each model.
}
\author{
Martin Vincent
}

