% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate.R
\name{parameters_stat.lsgl}
\alias{parameters_stat.lsgl}
\title{Extracting parameter statistics}
\usage{
\method{parameters_stat}{lsgl}(object, ...)
}
\arguments{
\item{object}{a lsgl object}

\item{...}{ignored}
}
\value{
a vector of length \code{nmod(x)} or a matrix containing the number of nonzero parameters of the models.
}
\description{
Extracts the number of nonzero parameters in each model.
}
\author{
Martin Vincent
}

