% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{lsgl.cv}
\alias{lsgl.cv}
\title{Deprecated cv function}
\usage{
lsgl.cv(x, y, intercept = TRUE, weights = NULL, grouping = NULL,
  groupWeights = NULL, parameterWeights = NULL, alpha = 1, lambda,
  d = 100, fold = 10L, cv.indices = list(), max.threads = NULL,
  use_parallel = FALSE, algorithm.config = lsgl.standard.config)
}
\description{
Deprecated cv function
}
\keyword{internal}

