% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lsgl.lambda}
\alias{lsgl.lambda}
\title{Deprecated lambda function}
\usage{
lsgl.lambda(x, y, intercept = TRUE, weights = NULL, grouping = NULL,
  groupWeights = NULL, parameterWeights = NULL, alpha = 1, d = 100L,
  lambda.min, lambda.min.rel = FALSE,
  algorithm.config = lsgl.standard.config)
}
\description{
Deprecated lambda function
}
\keyword{internal}

