% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{lsgl}
\alias{lsgl}
\title{Deprecated fit function}
\usage{
lsgl(x, y, intercept = TRUE, weights = NULL, grouping = NULL,
  groupWeights = NULL, parameterWeights = NULL, alpha = 1, lambda,
  d = 100, algorithm.config = lsgl.standard.config)
}
\description{
Deprecated fit function
}
\keyword{internal}

