% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{lsgl.standard.config}
\alias{lsgl.standard.config}
\title{Standard algorithm configuration}
\format{An object of class \code{list} of length 15.}
\usage{
lsgl.standard.config
}
\description{
\code{lsgl.standard.config <- lsgl.algorithm.config()}
}
\author{
Martin Vicnet
}
\keyword{datasets}

