% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling.R
\name{lsgl.subsampling}
\alias{lsgl.subsampling}
\title{Deprecated subsampling function}
\usage{
lsgl.subsampling(x, y, intercept = TRUE, weights = NULL, grouping = NULL,
  groupWeights = NULL, parameterWeights = NULL, alpha = 1, lambda,
  d = 100, train, test, collapse = FALSE, max.threads = NULL,
  use_parallel = FALSE, algorithm.config = lsgl.standard.config)
}
\description{
Deprecated subsampling function
}
\keyword{internal}

